src/requirements.txt
python-telegram-bot==20.0a0 # pre-release
